# [stdlib] Speed up bridged Dictionary instances