--- a/benchmark/CMakeLists.txt
+++ b/benchmark/CMakeLists.txt
@@ -65,6 +65,7 @@ set(SWIFT_BENCH_MODULES
     single-source/DictTest4
     single-source/DictTest4Legacy
     single-source/DictionaryBridge
+    single-source/DictionaryBridgeToObjC
     single-source/DictionaryCopy
     single-source/DictionaryGroup
     single-source/DictionaryKeysContains