`Dictionary`’s native storage classes and `_SwiftDeferredNSDictionary` override `-[NSDictionary getObjects:andKeys:]` instead of its safer replacement, `-[NSDictionary getObjects:andKeys:count:]`.

Overriding the correct method will considerably speed up some Cocoa operations on bridged dictionaries.

rdar://problem/39285882