@lorentey 2018-07-04
@swift-ci please test