@lorentey 2018-07-04
I don't think we have benchmarks for bridged dictionaries yet, but just in case:

@swift-ci please benchmark