@lorentey 2018-07-05
I added new microbenchmarks to help us track performance. Copying elements in bulk out of an `_SwiftDeferredNSDictionary` is now ~5 times faster:

<details open>
  <summary>Improvement (1)</summary>
  
TEST                              | OLD | NEW | DELTA  | SPEEDUP   
---                               | --- | --- | ---    | ---       
DictionaryBridgeToObjC_BulkAccess | 699 | 143 | -79.5% | **4.89x** 

</details>

<details open>
  <summary>No Changes (2)</summary>
  
TEST                              | OLD | NEW | DELTA  | SPEEDUP   
---                               | --- | --- | ---    | ---       
DictionaryBridgeToObjC_Access     | 774 | 760 | -1.8%  | 1.02x (?) 
DictionaryBridgeToObjC_Bridge     | 18  | 18  | +0.0%  | 1.00x     

</details>