@lorentey 2018-07-05
@swift-ci test